% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaANCOM.R
\name{adaANCOM}
\alias{adaANCOM}
\title{adaANCOM}
\usage{
adaANCOM(otu = NULL, Y = NULL, tree, tfun = t.test, smooth = 0.5, alpha = 0.05)
}
\arguments{
\item{otu}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{Y}{interested outcome.}

\item{tree}{the phylogenetic tree.}

\item{tfun}{the testing procedure, such as t.test or wilcox rank sum test for two-group comparison.}

\item{smooth}{the strategy for zero smooth. Can be choosen from a constant (e.g 0.5/1), or DM, ZIDM, MIX which corresponding the posterior mean or after likelihood ratio test to choose the best fitted model's.}

\item{alpha}{the significant level both for internal and leaf nodes.}
}
\value{
Returns a list:
\itemize{
\item V testing results for internal nodes V .
\item L testing results for leaf nodes L.
}
}
\description{
The main function for testing on the internal and leaf nodes. It first test on each internal node by
building log-ratios transformation and then test for each leaf node by build log-ratios transformation adaptively based on the first step.
}
\examples{
library(phyloseq)
data(COMBO)
otu.tab <- otu_table(COMBO)
tree <- phy_tree(COMBO)
group <- sample_data(COMBO)$bmicat1norm2ow3ob
group[group==3] <- 2
fit <- adaANCOM(otu=t(data.frame(otu.tab)), Y=group, tree=tree, tfun = t.test, smooth=0.5)
summary(fit)
}
