% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.dm.NR.R
\name{est.dm.NR}
\alias{est.dm.NR}
\title{Parameter estimation of Dirichlet-multinomial (DM) by Newton-Raphson (NR) algorithm.}
\usage{
est.dm.NR(X, init = NULL, iter = 1000, conv = 1e-06)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components.}

\item{init}{the initial value for the interative algorithm; if not given, we would like to use moment estimation.}

\item{iter}{the number of the interative steps.}

\item{conv}{the precison for stop the estimation.}
}
\value{
a list with the estimated parameters and log-likelihood value.
}
\description{
Given the data, we implemented NR to estimate parameters of DM which is faster than other methods.
}
