% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ascent.R
\name{find_ascent}
\alias{find_ascent}
\title{Find ascentors}
\usage{
find_ascent(edge, v)
}
\arguments{
\item{edge}{the edge from phylo object}

\item{v}{the number of the node}
}
\value{
all ascentors
}
\description{
Given the edge from phylo object and the number of the node, return its all ascentors
}
\examples{
library(phyloseq)
data(COMBO)
tree <- phy_tree(COMBO)
find_ascent(tree$edge, 1)
}
