% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_descent.R
\name{find_descent}
\alias{find_descent}
\title{Find descent}
\usage{
find_descent(edge, v)
}
\arguments{
\item{edge}{the edge from phylo object}

\item{v}{the number of the internal node}
}
\value{
all children nodes
}
\description{
Given the edge from phylo object and the number of the internal node, return its all descents.
}
\examples{
library(phyloseq)
data(COMBO)
tree <- phy_tree(COMBO)
find_descent(tree$edge, 55)
}
