% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiK_MN.R
\name{LiK_MN}
\alias{LiK_MN}
\title{Log-ikelihood ratio test (lrt) between Multinomial (MN) and Dirichlet-mulitnomial (DM) distribution.}
\usage{
LiK_MN(X, fdm = NULL)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components.}

\item{fdm}{the log-likelihood value of DM.}
}
\value{
p-value for the test.
}
\description{
Since MN nested in DM, we construct lrt between MN and DM, the degree of freedom equals to 1.
}
