% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikRT.R
\name{LikRT}
\alias{LikRT}
\title{Log-ikelihood ratio test (lrt) between Multinomial, Dirichlet-mulitnomial (DM), and Zero-inflate DM (ZIDM) distribution.}
\usage{
LikRT(X)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components.}
}
\value{
p-value for the test and the log-likelihood value
}
\description{
Since MN < DM < ZIDM,
}
\note{
also see \link{LiK_MN} or \link{LiK_DM}.
}
