% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikRTest.R
\name{LikRTest}
\alias{LikRTest}
\title{Log-ikelihood ratio test (lrt) between Multinomial, Dirichlet-mulitnomial (DM), and Zero-inflate DM (ZIDM) distribution.}
\usage{
LikRTest(X)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components.}
}
\value{
p-value for the test and the log-likelihood value
}
\description{
Since MN < DM < ZIDM,
}
\note{
also see \link{lrt_DM2ZIDM} or \link{lrt_MN2DM}.
}
