% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_dm.R
\name{loglik_dm}
\alias{loglik_dm}
\title{log-likelihood function for Dirichlet-multinomial (DM).}
\usage{
loglik_dm(X, alpha)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components.}

\item{alpha}{the corresponding parameters.}
}
\value{
the log-likelihood value for DM.
}
\description{
Given data X and the parameters, we could calcualte the log-likelihood value of them.
}
