% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_zidm.R
\name{loglik_zidm}
\alias{loglik_zidm}
\title{log-likelihood function for Zero-inflated Dirichlet-multinomial (ZIDM) distribution.}
\usage{
loglik_zidm(X, alpha, pi)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components}

\item{alpha}{the corresponding parameters for Dirichlet-multinomial part.}

\item{pi}{the correpsonding parameters for Zero-inflated part}
}
\value{
the log-likelihood value for ZIDM
}
\description{
Given data X and the parameters, we could calcualte the log-likelihood value of them.
}
