% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_DM2ZIDM.R
\name{lrt_DM2ZIDM}
\alias{lrt_DM2ZIDM}
\title{Log-ikelihood ratio test (lrt) between Dirichlet-mulitnomial (DM) and Zero-inflate DM (ZIDM) distribution.}
\usage{
lrt_DM2ZIDM(X, fdm = NULL, fzidm = NULL)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components.}

\item{fdm}{the loglikelihood value of DM.}

\item{fzidm}{the loglikelihood value of ZIDM.}
}
\value{
p-value for the test.
}
\description{
Since DM nested in ZIDM, we construct lrt between them, the degree of freedom equals to the number of the conmponent minus 1.
}
