% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.dm.R
\name{mom.dm}
\alias{mom.dm}
\title{Moment estimation for Dirichlet-multinomial}
\usage{
mom.dm(X)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components}
}
\value{
a estimated vector
}
\description{
Moment estimation for Dirichlet-multinomial.
}
