% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outliers.R
\name{Outliers}
\alias{Outliers}
\title{Outlier detection}
\usage{
Outliers(X, P)
}
\arguments{
\item{X}{a data frame or matrix containing the count data, with two columns.}

\item{P}{a data frame or matrix containing the transformed data, with two columns.}
}
\value{
a list containing the index of the outliers and the value of the log-ratios both for noraml and abnormal value.
}
\description{
To avoid abnormal value based on the transformed data of the log-ratios, we use the maximum of the log-ratios of non-zeros as the threshold.
}
\examples{

# data generation
N <- 100
D1 <- D2 <- round(runif(N, 1, 100))
Pi <- rzidirichlet(N=N, alpha =  c(3, 6), pi=0.1)
which(Pi[,1]==0) # true zero
X1 <- matrix(NA, N, 2)
for(i in 1:N) X1[i, ] <- rmultinom(1, D1[i], prob=Pi[i, ])

prop <- Smooth_X(X1, type='MIX', rel=TRUE, outlier=FALSE)

res <- Outliers(X1, prop)
res
}
