% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorMeanTrans.R
\name{PosteriorMeanTrans}
\alias{PosteriorMeanTrans}
\title{Posterior mean transformation}
\usage{
PosteriorMeanTrans(otu, tree, model = "MIX")
}
\arguments{
\item{otu}{the OTU count table.}

\item{tree}{the phylogenetic tree.}

\item{model}{smooth methods, also see \link{Smooth_X}.}
}
\value{
Returns two lists:
\itemize{
\item count count abundance for internal nodes V, leaf nodes L and count on each subtree Y.
\item probability abundance for internal nodes V, leaf nodes L and count on each subtree Y.
}
}
\description{
Data transformation based on the posterior mean.
}
