% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powfdr.R
\name{powfdr}
\alias{powfdr}
\title{Evaluation of the model.}
\usage{
powfdr(p.adj, dif = NULL, a = 0.05)
}
\arguments{
\item{p.adj}{the adjusted p-value or statistics.}

\item{dif}{the true differential conmonent.}

\item{a}{the significant level.}
}
\value{
a vector consist of recall, precision, and F1 score squentially.
}
\description{
Given the statistics and threshold, we get the positive results.
Then compared to the true signal, we could calculate the recall, precision, and F1 score.
}
