% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzidirichlet.R
\name{rzidirichlet}
\alias{rzidirichlet}
\title{Zero-inflated Dirichlet (ZID) distribution.}
\usage{
rzidirichlet(N = 10, alpha, pi = 0.1, seed = 1)
}
\arguments{
\item{N}{a value for the sample size.}

\item{alpha}{a vector for the with K positive parameters.}

\item{pi}{a value or vector with length K-1, for the zero proportion of the first K-1 componenets.}

\item{seed}{a value for the seeds.}
}
\value{
a probability matrix, with some zero.
}
\description{
Generate ZID probability with given parameters, sample size, it will degenerate to zero-inflated beta with only two componenets.
}
\examples{
X <- rzidirichlet(N=10, alpha=rep(1, 5), pi=0.1)
X
}
