% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth_Tree.R
\name{Smooth_Tree}
\alias{Smooth_Tree}
\title{Zero smooth on the tree.}
\usage{
Smooth_Tree(Yotu, tree, type = "DM", rel = F, outlier = F)
}
\arguments{
\item{Yotu}{a list which containing all data matrix on each internal nodes.}

\item{tree}{the phylogenetic tree.}

\item{type}{smooth methods, also see \link{Smooth_X}.}

\item{rel}{If true, the output will be the posterior; otherwise, we product the posterior by the sample depth to output the smoothed count.}

\item{outlier}{If true, remove the outliers, otherwise keep them.}
}
\value{
a list, one is the smoothed data on each internal nodes and other one is a matrix which consist of all internal and leaf nodes.
}
\description{
Smooth on each internal node indenpendently.
}
