% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ytree2otu.R
\name{Ytree2otu}
\alias{Ytree2otu}
\title{Transform the count on each internal node to the leafs.}
\usage{
Ytree2otu(Yotu, tree)
}
\arguments{
\item{Yotu}{a list, containing each count matrix on the subtree.}

\item{tree}{the phylogenetic tree.}
}
\description{
Given a list which consist of all internal counts, we extract the count of the leaves.
}
\note{
also see \link{Ytreefun}.
}
