% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ytreefun.R
\name{Ytreefun}
\alias{Ytreefun}
\title{Calculate the count on each subtree.}
\usage{
Ytreefun(otu.tab, tree)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{tree}{the phylogenetic tree.}
}
\value{
Returns a list, containing each count matrix on the subtree.
}
\description{
Follow the additive foumla, each internal counts equals to summation of his children's.
}
\examples{
library(phyloseq)
data(COMBO)
otu.tab <- t(otu_table(COMBO))
tree <- phy_tree(COMBO)
ytree <- Ytreefun(otu.tab, tree)
length(ytree)
}
