% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth_X.R
\name{Smooth_X}
\alias{Smooth_X}
\title{Zero smooth.}
\usage{
Smooth_X(X, type = "DM", rel = F, outlier = F)
}
\arguments{
\item{X}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the components}

\item{type}{the smoothing methods which could be specified including 0.5, DM, ZIDM, or MIX which using lrt to choose the best fitted one.}

\item{rel}{If true, the output will be the posterior; otherwise, we product the posterior by the sample depth to output the smoothed count.}

\item{outlier}{If true, remove the outliers, otherwise keep them.}
}
\value{
smoothed data matrix
}
\description{
We used the posterior mean transfromation to smooth zero. If the prior is not given,
we could use likelihood ratio test to chooce the best fitted model and the corresponding posterior mean to smooth zeros.
}
\examples{

# data generation
N <- 100
D1 <- D2 <- round(runif(N, 1, 100))
Pi <- rzidirichlet(N=N, alpha =  c(3, 6), pi=0.1)
which(Pi[,1]==0) # true zero
X1 <- matrix(NA, N, 2)
for(i in 1:N) X1[i, ] <- rmultinom(1, D1[i], prob=Pi[i, ])

prop <- Smooth_X(X1, type='MIX', rel=TRUE, outlier=FALSE)

res <- Outliers(X1, prop)
res

}
